- removed web export from index - fix package.json - added local scope to browser.js - export server code inside the root file
